# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REST API - Authentication & Roles",
        "description": "API REST desarrollada con Laravel 12 que implementa autenticación con Sanctum, autorización por roles y permisos, y operaciones CRUD completas para productos.",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor Docker"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar productos",
                "description": "Obtiene una lista paginada de productos. Permite filtrar por categoría.",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filtrar productos por categoría",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Electronics"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Laptop Pro 15"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "High-performance laptop..."
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "1299.99"
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Electronics"
                                                    },
                                                    "created_by": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Admin User"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "admin@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-12-30 22:10:45"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-12-30 22:10:45"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto. Requiere el permiso 'create-products'.",
                "operationId": "eddf8c7847239ddc1040daadd814e21b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del producto",
                                        "type": "string",
                                        "example": "New Product"
                                    },
                                    "description": {
                                        "description": "Descripción del producto (opcional)",
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "price": {
                                        "description": "Precio del producto",
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "stock": {
                                        "description": "Cantidad en stock",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "category": {
                                        "description": "Categoría del producto (opcional)",
                                        "type": "string",
                                        "example": "Electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New Product"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Product description"
                                        },
                                        "price": {
                                            "type": "string",
                                            "example": "99.99"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Electronics"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to create products."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingrese el token de autenticación obtenido del endpoint /login o /register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}